+++ Civi/ActionProvider/Action/Contribution/ContributionRecurRepeatTransaction.php
@@ -1,0 +1,65 @@
+<?php
+/**
+ * @author Jaap Jansma <jaap.jansma@civicoop.org>
+ * @license AGPL-3.0
+ */
+
+namespace Civi\ActionProvider\Action\Contribution;
+
+use Civi\ActionProvider\Action\AbstractAction;
+use Civi\ActionProvider\ConfigContainer;
+use Civi\ActionProvider\Parameter\OptionGroupSpecification;
+use Civi\ActionProvider\Parameter\ParameterBag;
+use Civi\ActionProvider\Parameter\ParameterBagInterface;
+
+use Civi\ActionProvider\Parameter\Specification;
+use Civi\ActionProvider\Parameter\SpecificationBag;
+use Civi\ActionProvider\Utils\CustomField;
+use CRM_ActionProvider_ExtensionUtil as E;
+
+class ContributionRecurRepeatTransaction extends AbstractAction {
+
+  /**
+   * Run the action
+   *
+   * @param ParameterBagInterface $parameters
+   *   The parameters to this action.
+   * @param ParameterBagInterface $output
+   * 	 The parameters this action can send back
+   * @return void
+   * @throws \Exception
+   */
+  protected function doAction(ParameterBagInterface $parameters, ParameterBagInterface $output) {
+    $contribution_params['contribution_recur_id'] = $this->configuration->getParameter('contribution_recur_id');
+    $result = civicrm_api3('Contribution', 'repeattransaction', $contribution_params);
+  }
+
+  /**
+   * @return \Civi\ActionProvider\Parameter\SpecificationBag
+   */
+  public function getOutputSpecification() {
+    return new SpecificationBag(array());
+  }
+
+  /**
+   * Returns the specification of the configuration options for the actual action.
+   *
+   * @return SpecificationBag
+   */
+  public function getConfigurationSpecification() {
+    return new SpecificationBag(array());
+  }
+
+  /**
+   * Returns the specification of the parameters of the actual action.
+   *
+   * @return SpecificationBag
+   */
+  public function getParameterSpecification() {
+    $specs = new SpecificationBag(array(
+      new Specification('contribution_recur_id', 'Integer', E::ts('Contribution Recur ID'), false),
+    ));
+    return $specs;
+  }
+
+}